// src/shared/types.ts
export enum Events {
	AUTHENTICATION = 'authentication',
	SEND_MESSAGE = 'send_message',
	GET_MESSAGES = 'get_messages',
	UPDATE_READ = 'update_read',
	READ_MESSAGE = 'read_message',
	UPDATE_ONLINE = 'update_online',
	NEW_CHAT = 'new_chat',
	DELETE_CHAT = 'delete_chat',
	USER_LEAVE = 'user_leave',
	ADD_USER = 'add_user',
	SET_TYPING_STATUS = 'set_typing_status',
	CHANGE_GROUP_COLOR = 'change_group_color',
	ARCHIVE = 'archive',
	NOTIFICATIONS = 'notifications_status',
	DELETE_MESSAGE = 'delete_message'
}